[refactor] Service 인터페이스 분리
